[testsuite] [ppc64] expect error on vxworks too
These ppc lp64 tests check for errors or warnings on -mno-powerpc64.
On powerpc64-*-vxworks* we get the same errors as on most other
covered platforms, but the tests did not mark them as expected for
this target.  On powerpc-*-vxworks*, the tests are skipped because
lp64 is not satisfied, so I'm naming powerpc*-*-vxworks* rather than
something more specific.


for  gcc/testsuite/ChangeLog

	* gcc.target/powerpc/pr106680-1.c: Error on vxworks too.
	* gcc.target/powerpc/pr106680-2.c: Likewise.
	* gcc.target/powerpc/pr106680-3.c: Likewise.