[testsuite] require sqrt_insn effective target where needed
Some tests fail on ppc and ppc64 when testing a compiler [with options
for] for a CPU [emulator] that doesn't support the sqrt insn.

The gcc.dg/cdce3.c is one in which the expected shrink-wrap
optimization only takes place when the target CPU supports a sqrt
insn.

The gcc.target/powerpc/pr46728-1[0-4].c tests use -mpowerpc-gpopt and
call sqrt(), which involves the sqrt insn that the target CPU under
test may not support.

Require a sqrt_insn effective target for all the affected tests.


for  gcc/testsuite/ChangeLog

	* gcc.dg/cdce3.c: Require sqrt_insn effective target.
	* gcc.target/powerpc/pr46728-10.c: Likewise.  Drop darwin
	explicit skipping.
	* gcc.target/powerpc/pr46728-11.c: Likewise.  Likewise.
	* gcc.target/powerpc/pr46728-13.c: Likewise.  Likewise.
	* gcc.target/powerpc/pr46728-14.c: Likewise.  Likewise.