PHIOPT: Value-replacement check undef
While moving value replacement part of PHIOPT over
to use match-and-simplify, I ran into the case where
we would have an undef use that was conditional become
unconditional. This prevents that. I can't remember at this
point what the testcase was though.

Bootstrapped and tested on x86_64-linux-gnu with no regressions.

gcc/ChangeLog:

	* tree-ssa-phiopt.cc (value_replacement): Reject undef variables
	so they don't become unconditional used.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>