testsuite: gm2: Remove timeout overrides [PR114886]
A large number of gm2 tests are timing out even on current Solaris/SPARC
systems.  As detailed in the PR, the problem is that the gm2 testsuite
artificially lowers many timeouts way below the DejaGnu default of 300
seconds, often as short as 10 seconds.  The problem lies both in the
values (they may be appropriate for some targets, but too low for
others, especially under high load) and the fact that it uses absolute
values, overriding e.g. settings from a build-wide site.exp.

Therefore this patch removes all those overrides, restoring the
defaults.

Tested on sparc-sun-solaris2.11 (where all the previous timeouts are
gone) and i386-pc-solaris2.11.

2024-04-29  Rainer Orth  <ro@CeBiTec.Uni-Bielefeld.DE>

	gcc/testsuite:
	PR modula2/114886
	* lib/gm2.exp: Don't load timeout-dg.exp.
	Don't set gm2_previous_timeout.
	Don't call dg-timeout.
	(gm2_push_timeout, gm2_pop_timeout): Remove.
	(gm2_init): Don't call dg-timeout.
	* lib/gm2-torture.exp: Don't load timeout-dg.exp.
	Don't set gm2_previous_timeout.
	Don't call dg-timeout.
	(gm2_push_timeout, gm2_pop_timeout): Remove.

	* gm2/coroutines/pim/run/pass/coroutines-pim-run-pass.exp: Don't
	load timeout-dg.exp.
	Don't call gm2_push_timeout, gm2_pop_timeout.
	* gm2/examples/map/pass/examples-map-pass.exp: Don't call
	gm2_push_timeout, gm2_pop_timeout.
	* gm2/iso/run/pass/iso-run-pass.exp: Don't load timeout-dg.exp.
	Don't call gm2_push_timeout, gm2_pop_timeout.
	* gm2/pimlib/base/run/pass/pimlib-base-run-pass.exp: Don't load
	timeout-dg.exp.
	Don't call gm2_push_timeout, gm2_pop_timeout.
	* gm2/projects/iso/run/pass/halma/projects-iso-run-pass-halma.exp:
	Don't call gm2_push_timeout, gm2_pop_timeout.
	* gm2/switches/whole-program/pass/run/switches-whole-program-pass-run.exp:
	Don't load timeout-dg.exp.
	Don't call gm2_push_timeout, gm2_pop_timeout.